+++ src/main/java/blackjack/view/OutputView.java
@@ -2,6 +2,7 @@ package blackjack.view;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Participant;
 import blackjack.domain.participant.Player;
 import blackjack.domain.participant.Players;
 import java.util.List;
@@ -15,7 +16,7 @@ public class OutputView {
         System.out.println();
         System.out.println("딜러와 " + readPlayers.stream()
                 .map(player -> player.getName().getValue())
-                .collect(Collectors.joining(", ")) + "에게 2장을 나누었습니다." );
+                .collect(Collectors.joining(", ")) + "에게 2장을 나누었습니다.");
 
         System.out.println("딜러: " + dealer.showOneCard());
 
@@ -25,10 +26,10 @@ public class OutputView {
         System.out.println();
     }
 
-    public static void printCardHandStatus(Player player) {
-        System.out.printf("%s카드: %s%n",
-                player.getName().getValue(),
-                player.getCards().getCardHand().stream()
+    public static String showCardHandStatus(Participant participant) {
+        return String.format("%s 카드: %s",
+                participant.getName().getValue(),
+                participant.getCards().getCardHand().stream()
                         .map(Card::toString)
                         .collect(Collectors.joining(", ")));
     }
@@ -36,5 +37,16 @@ public class OutputView {
     public static void printDealerStatus() {
         System.out.println();
         System.out.println("딜러는 16이하라 한장의 카드를 더 받았습니다.");
+    }
+
+    public static void printCardHandStatus(Participant participant) {
+        System.out.println(showCardHandStatus(participant));
+    }
+
+    public static void printFinalStatus(Dealer dealer, Players players) {
+        System.out.println(showCardHandStatus(dealer) + " 결과 - " + dealer.calculateBestScore());
+        for (Player readPlayer : players.getPlayers()) {
+            System.out.println(showCardHandStatus(readPlayer) + " 결과 - " + readPlayer.calculateBestScore());
+        }
     }
 }
